requirements.txt
soundcard
numpy
dearpygui
matplotlib
pillow